refactor: reuse test setup for data sources
